Refactor getQuestPayments
Rewrite it to make more use of async/await to make it (imho) more readable.
Also "clone" the item via the spread operator instead of modifying it. Modifing data in hooks is an anti-pattern in react, so while I've been in that function ;)
And lastly the fix: await the result of `getPrizeTransaction` to set it directly and make react render the whole result.

Note: this patch increases the time until the badges are rendered correctly as react needs to wait until the prizes are loaded.
I do this in parallel already to reduce the time as much as possible.